+++ src/background/modules/mod.rs
@@ -1,3 +1,5 @@
 pub mod power;
 pub mod file_extensions;
 pub mod cli;
+pub mod input;
+pub mod tray;