+++ src/background/seelen_weg/icon_extractor.rs
@@ -2,6 +2,7 @@ use color_eyre::eyre::eyre;
 use image::ImageBuffer;
 use image::RgbaImage;
 use itertools::Itertools;
+use tauri::AppHandle;
 use widestring::U16CString;
 use windows::core::PCWSTR;
 use windows::Win32::Graphics::Gdi::CreateCompatibleDC;
@@ -24,8 +25,11 @@ use std::arch::x86_64::_mm_setr_epi8;
 #[cfg(target_arch = "x86_64")]
 use std::arch::x86_64::_mm_shuffle_epi8;
 use std::arch::x86_64::_mm_storeu_si128;
+use std::path::PathBuf;
 
 use crate::error_handler::Result;
+use crate::utils::app_data_path;
+use crate::utils::filename_from_path;
 
 /// Convert BGRA to RGBA
 ///
@@ -151,4 +155,36 @@ pub fn convert_hicon_to_rgba_image(hicon: &HICON) -> Result<RgbaImage> {
             .expect("Failed to create image buffer");
         return Ok(image);
     }
+}
+
+/// returns the path of the icon extracted from the executable or copied if is an UWP app.
+///
+/// If the icon already exists, it returns the path instead overriding, this is needed for allow user custom icons.
+pub fn extract_and_save_icon(handle: &AppHandle, exe_path: &str) -> Result<PathBuf> {
+    let gen_icons_paths = app_data_path(handle).join("icons");
+    if !gen_icons_paths.exists() {
+        std::fs::create_dir_all(&gen_icons_paths)?;
+    }
+
+    let filename = filename_from_path(exe_path);
+    let icon_path = gen_icons_paths.join(filename.replace(".exe", ".png"));
+    let icon_path_uwp = gen_icons_paths.join(filename.replace(".exe", "_uwp.png"));
+
+    if icon_path_uwp.exists() {
+        return Ok(icon_path_uwp);
+    }
+
+    if icon_path.exists() {
+        return Ok(icon_path);
+    }
+
+    let images = get_images_from_exe(exe_path);
+    if let Ok(images) = images {
+        // icon on index 0 always is the app showed icon
+        if let Some(icon) = images.get(0) {
+            icon.save(&icon_path)?;
+        }
+    }
+
+    Ok(icon_path)
 }